% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_packages}
\alias{check_packages}
\title{Check for packages}
\usage{
check_packages(package_list, message = NULL)
}
\arguments{
\item{package_list}{A list of packages to check}

\item{message}{A specialized message to print. Otherwise a generic
message will be printed}
}
\value{
None
}
\description{
Checks to ensure all packages required for a function are installed. If
any packages aren't installed, the function will quit and print an
error message listing packages that need to be installed.
}
