% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{Convert between human and mouse}
\usage{
convertHumanGeneList(x, convert = "human_mouse")
}
\arguments{
\item{x}{The list of genes to convert}

\item{convert}{The type of conversion to do. Can be "human_mouse"
to convert human genes to mouse genes or "mouse_human" to convert
mouse genes to human}
}
\value{
a dataframe containing both human and mouse genes for the
input vector
}
\description{
Basic function to convert human to mouse gene name or mouse to
human using biomart
}
\examples{
\dontrun{
mouse_human_list <- convertHumanGeneList(x = c("INS1", "INS2", "GCG",
                                               "ARX", "NKX2-2", "SST"),
                                         convert = "human_mouse")
mouse_human_list <- convertHumanGeneList(x = c("Ins1", "Ins2", "Gcg",
                                               "Arx", "Nkx2-2", "Sst"),
                                         convert = "mouse_human")
}
}
