% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{create_spatial_seurat}
\alias{create_spatial_seurat}
\title{Creates a spatial seurat object}
\usage{
create_spatial_seurat(
  results_dir,
  sample_name,
  filter_matrix = TRUE,
  filename = "filtered_feature_bc_matrix.h5"
)
}
\arguments{
\item{results_dir}{The path to the results directory. The structure should be
results_dir/sample_name/outs}

\item{sample_name}{The name of the sample. This helps find the input data. The
path should be results_dir/sample_name/outs}

\item{filter_matrix}{OPTIONAL If the matrix should be filtered to only include
spots determined to be over tissue. Used by Read10X_Image.}

\item{filename}{OPTIONAL The name of the h5 file to load. Default is
filtered_feature_bc_matrix.h5 which should generally work.}
}
\value{
A seurat object with image data loaded.
}
\description{
This function will create a seurat object. It needs the path to the directory
containing output from cellranger count that also includes the images.
}
\examples{
\dontrun{
seurat_object <- create_spatial_seurat(sample     = "/home/users/kwells/experiment_1",
                                       count_path = "WT_1")
}
}
