% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{de_to_pathview}
\alias{de_to_pathview}
\title{Run differential expression followed by pathview.}
\usage{
de_to_pathview(
  path_id_list,
  seurat_object = NULL,
  seurat_de = NULL,
  out_dir = "pathview",
  pval = 0.05,
  gen_id = "mmu",
  ...
)
}
\arguments{
\item{path_id_list}{A named list of KEGG path ids. The ids are the 5 number
ids, the name will be associated with the output file}

\item{seurat_object}{OPTIONAL A seurat object that has already been
normalized. Must provide either a seurat object or a seurat de.}

\item{seurat_de}{OPTIONAL the returned object from find_markers. If not
provided, find markers will be run. Must provide either a seurat object or
a seurat de.}

\item{out_dir}{OPTIONAL an output directory for the pathview plots. Default
it "pathview"}

\item{pval}{OPTIONAL the pvalue cutoff for the adjusted pvalue from the DE
test. Default is 0.05.}

\item{...}{OPTIONAL arguments supplied to find_markers}

\item{seurat_assay}{OPTIONAL which assay to pull the data from when
performing DE. Default is "RNA"}
}
\value{
Nothing is returned, instead pathview plots are made and saved in
the specified directory
}
\description{
This function allows you to run pathview given a matrix of DE output. It can
also run pathview on precomputed DE output.
}
\examples{
\dontrun{
de_to_pathview(path_id_list  = c(NFKB_sig_path = "04064", mapk_signaling = "04010"),
               seurat_object = seurat_object))
de_to_pathview(path_id_list = c(NFKB_sig_path = "04064", mapk_signaling = "04010"),
               seurat_de    = de_output))
}
}
