% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{featDistPlot}
\alias{featDistPlot}
\title{Plot a group of violin or jitter plots}
\usage{
featDistPlot(
  seurat_object,
  geneset,
  cell_cycle = FALSE,
  plot_type = "violin",
  color = NULL,
  sep_by = "cluster",
  save_plot = NULL,
  col_by = NULL,
  plot_median = TRUE,
  combine = TRUE,
  dodge = 1,
  width = 0.9,
  assay = NULL,
  ...
)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{geneset}{What genes to use as the y axis in the plot. Can be many, but I
recommend not going above 3 because all plots will be returned as one object.}

\item{cell_cycle}{OPTIONAL Only an option if making a jitter plot. Can color each
cel by the cell cycle phase. I originally did this so the colors would be consistent.
Default is FALSE. If set to TRUE, it will override your col_by and color arguments.}

\item{plot_type}{OPTIONAL The type of plot. The options are "violin", "jitter", or
"both". "jitter" will make only a jitter plot, "violin" will make only a violin plot
and "both" will make a violin plot with jitter dots overlayed on tope.}

\item{color}{The color palette used to color either the points for a jitter plot or
the violins. Default is Set1 from RColorBrewer}

\item{sep_by}{OPTIONAL what to use to separate the jitter or violins on the x axis.
Can be any discrete value from the metadata. Deafult is "cluster"}

\item{save_plot}{OPTIONAL A path to a file name if you want the plot saved. It
will be saved as a PDF. If nothing is provided, the plot will be returned but not
saved}

\item{col_by}{OPTIONAL what to use to color the cells (or violins). Generally for a violin
plot this is the same as sep_by, but for jitter, it can be any discrete column from meta
data. Default is NULL.}

\item{plot_median}{OPTIONAL if the median value should be included in the
violin plot. Default is TRUE}

\item{combine}{OPTIONAL if the final plots should be made into a figure or
returned as a list. Default (TRUE) returns a figure.}

\item{dodge}{OPTIONAL how to adjust the placement of the violin plot. Default is 1}

\item{width}{OPTIONAL how to adjust how wide the violin plots are. Default is 0.9}

\item{assay}{OPTIONAL what assay to pull the data from. Default is NULL.}

\item{...}{arguments passed to other functions}
}
\value{
Either a list of plots (if combine = FALSE) or a gridExtra object with all
plots stacked (if combine = TRUE)
}
\description{
This function will plot a violin or jitter plot of the expression of a gene from
your seurat object. The plot could also be of some continuous data from your
meta data. The violin plot can be separated by any discrete variable in your
meta data. If multiple genes are supplied, a single plot conisiting of all violin
or jitter plots will be returned with one plot per row.
}
\examples{
\dontrun{
featDistPlot(seurat_object = seurat_object,
             geneset       = "INS1",
             col_by        = "cluster",
             sep_by        = "cluster",
             color         = c(1 = "#FFFFFF", 2 = "#FF0000"))
featDistPlot(seurat_object = seurat_object,
             geneset       = "INS1",
             col_by        = "cluster",
             sep_by        = "cluster",
             color         = c(1 = "#FFFFFF", 2 = "#FF0000"),
             plot_type     = "both")
featDistPlot(seurat_object = seurat_object,
             geneset       = c("INS1", "INS2", "GCG"),
             col_by        = "cluster",
             sep_by        = "cluster",
             color         = c(1 = "#FFFFFF", 2 = "#FF0000"),
             combine       = FALSE)
}
}
