% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{find_markers}
\alias{find_markers}
\title{Run differential expression}
\usage{
find_markers(
  seurat_object,
  test_idents = NULL,
  seurat_assay = "RNA",
  group = NULL,
  group_ident_1 = NULL,
  group_ident_2 = NULL
)
}
\arguments{
\item{seurat_object}{A seurat object that has already been normalized.}

\item{test_idents}{OPTIONAL The identities from the metadata used to perform
DE. If no identities are provided, the identites already set in the object
will be used.}

\item{seurat_assay}{OPTIONAL which assay to pull the data from when
performing DE. Default is "RNA"}

\item{group}{OPTIONAL to be used when running DE between different samples.
The group refers to the column contining the sample info. Default is NULL.}

\item{group_ident_1}{OPTIONAL The sample to use as the main sample to test
against. Defaults to the first value of the factor}

\item{group_ident_2}{OPTIONAL which sample to use as the second group
if more than 2 samples are in the seurat object.}
}
\value{
A dataframe containing all markers.
}
\description{
This function runs differential expression analysis using Seurat's
functions. This is called by several functions but can be used on
its own as well. Can run a pairwise DE or a one vs all DE.
}
\examples{
\dontrun{
marker_df <- find_markers(seurat_object = seurat_object)
marker_df <- find_markers(seurat_object = seurat_object,
                          group         = "cluster")
marker_df <- find_markers(seurat_object = seurat_object,
                          group         = "cluster",
                          group_ident_1 = 1)
}
}
