% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{find_write_markers_pairwise}
\alias{find_write_markers_pairwise}
\title{Find and write markers}
\usage{
find_write_markers_pairwise(
  seurat_object,
  save_dir,
  meta_col = "RNA_cluster",
  assay = "RNA",
  pval = 0.05,
  logfc = 0.5,
  gene_lists = NULL
)
}
\arguments{
\item{seurat_object}{A seurat object that has already been normalized.}

\item{save_dir}{The path to save all of the output file.}

\item{meta_col}{OPTIONAL The column in the meta data to use to identify
markers. Default is RNA_cluster.}

\item{assay}{OPTIONAL the assay to run the differential expression.
Default is RNA.}

\item{logfc}{OPTIONAL The cutoff for the log fold change for returned
genes. Default is 0.5.}

\item{gene_lists}{OPTIONAL gene lists to compare against differentially
expressed genes. Can provide many vectors of genes as a list. Provided
lists will be used to run a hypergeometric test against differentially
expressed genes identified from each cluster. If no gene lists are
included, no hypergeometric tests will be run.}

\item{p_val}{OPTIONAL The cutoff for p value for returned genes.
Default is 0.05}
}
\value{
A dataframe containing all markers. It will also write a csv file
with the DE genes, a csv file with the hypergeometric test output (if run)
and an excel file with both hypergeometric output and DE genes.
}
\description{
This function finda all markers of all clusters for a specific column
in your seurat metadata by running FindMarkers in a pairwise fashion.
It writes these markers to both a csv file and an excel document with
one sheet per group from the meta data (ex 1 sheet per cluster). This
function will further run and write the output of hypergeometric tests
on supplied gene lists using the identified markers. It will perform a
p-value fdr correction based on the number of lists supplied. This
function is meant to be called by find_write_markers and not used directly.
}
