% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_seurat_assay}
\alias{get_seurat_assay}
\title{Get Seurat Assay}
\usage{
get_seurat_assay(object, type = "counts", assay = "RNA")
}
\arguments{
\item{object}{A seurat object}

\item{type}{What type should be returned, data or counts}

\item{assay}{What assay the data should be pulled from}
}
\value{
The resulting data matrix.
}
\description{
Determines the seurat version and grabs the assay data using
the appropriate arguments.
}
