% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{group_cells}
\alias{group_cells}
\title{Run UMAP and clustering}
\usage{
group_cells(
  sample_object,
  sample_name = NULL,
  save_dir = NULL,
  nPCs = 10,
  resolution = 0.8,
  assay = "RNA",
  HTO = FALSE,
  reduction = NULL,
  ...
)
}
\arguments{
\item{sample_object}{A seurat object}

\item{sample_name}{The name of the sample. This will be used for naming output
plots}

\item{save_dir}{Where to save plots that are generated (UMAPs of cluster and
original identity, and HTO identity if incldued)}

\item{nPCs}{OPTIONAL How many PCs to use to generate the UMAP and clustering. While
a default is provided, it is highly recommeded that you tune this to your own data.
Default is 10.}

\item{resolution}{OPTIONAL What resolution to use for clustering. Passed to FindClusters
A default is provided that tends to be a good starting point, but you are encouraged to
find the resolution appropriate to your dataset. Default is 0.8.}

\item{assay}{OPTIONAL What assay to use. This is used to locate the PCA reduction
(the reduction name provided assumes that no reduction name was supplied to PCA_dimRed,
if you did provide a different reduction name, you must use "reduction" below).
This is also used to name the output plots, reductions, and clusters. Can be "RNA",
"ADT", "SCT", or "integrated" Default is "RNA".}

\item{HTO}{OPTIONAL if HTOs were included in the seurat object. Default is FALSE}

\item{reduction}{OPTIONAL the name of the PCA reduction. Not required if you used
the default reduction names in PCA_dimRed. Default is NULL}

\item{...}{other arguments passed to plotDimRed}
}
\value{
A seurat object with new cluster information in the metadata (RNA_cluster
if assay = "RNA", SCT_cluster if assay = "SCT", integrated_cluster if assay =
"integrated", and ADT_cluster if assay = "ADT"), and a new UMAP reduction named
by the reduction you started with.
}
\description{
This function will run both UMAP and clustering on a seurat object.
}
\examples{
umap_res <- group_cells(sample_object = splen_so)
umap_res <- group_cells(sample_object = splen_so,
                             nPCs = 25,
                             resolution = 1.2)
}
