% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{groupDiscretePlots}
\alias{groupDiscretePlots}
\title{Plot one dimensionality reduction colored by discrete data}
\usage{
groupDiscretePlots(
  group,
  plot_df,
  col_by,
  axis_names = c("dim1", "dim2"),
  color = NULL,
  show_legend = TRUE,
  size = 0.25,
  ggrastr = FALSE,
  raster_scale = 1,
  raster_res = 300
)
}
\arguments{
\item{group}{The group to color.}

\item{plot_df}{A dataframe created by plotDimRedSingle}

\item{col_by}{What to color the plot by. May be a gene, adt, or column
from the metadata. It can either be discrete or continuous.}

\item{axis_names}{OPTIONAL What will be supplied to label the x and y axis}

\item{color}{OPTIONAL what colors to use to color the plot. Default is Set1
from RColorBrewer}

\item{show_legend}{OPTIONAL if a legend should be shown on the final plot. Deault is
TRUE}

\item{size}{OPTIONAL The size of the points. Default is 0.25}

\item{ggrastr}{OPTINAL If the plot should be rastarized. This is mostly helpful
for large datasets. All of the points will be a png while the rest will still
be editable. Default is FALSE (don't rasterize the plot)}

\item{raster_scale}{OPTIONAL The scale to use. Can be helpful if the rasterized
plot looks fuzzy. Default is 1}

\item{raster_res}{OPTIONAL. Can be helpful to change if the rasterized plot
looks fuzzy. Default is 300.}
}
\description{
This function is meant to be called by plotDimRed and not used on it's own!!!
}
\keyword{internal}
