% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{hypergeometric_test}
\alias{hypergeometric_test}
\title{Runs a hypergeometric test}
\usage{
hypergeometric_test(
  seurat_object,
  gene_list,
  DE_table,
  DE_p_cutoff = 0.05,
  DE_lfc_cutoff = 0.5,
  correction_method = "fdr"
)
}
\arguments{
\item{seurat_object}{A seurat object that has already been normalized.}

\item{gene_list}{gene lists to compare against differentially
expressed genes. Can provide many vectors of genes as a list. Provided
lists will be used to run a hypergeometric test against differentially
expressed genes identified from each cluster.}

\item{DE_table}{A data frame with differential expression results that can
be made by running FindAllMarkers}

\item{DE_p_cutoff}{OPTIONAL The cutoff for p value for differentially expressed
genes to be used in the hypergeometric test. Default is 0.05}

\item{DE_lfc_cutoff}{OPTIONAL The cutoff for the log fold change for differentially
expressed genes to be used in the hypergeometric test. Default is 0.5.}

\item{correction_method}{OPTIONAL The method used to correct the p-values for
multiple correction. Default is "fdr". Can be any method used by p.adjust.}
}
\value{
A dataframe containing hypergeometric output. Includes the expected number
of genes from each set seen, the enrichment of the gene set, a p-value, and a
corrected p-value for each cluster and each gene set.
}
\description{
Given a seurat object, results from a DE test, and a set of gene lists,
this runs a hypergeometric test (fishers exact test) to identify significant
enrichment of the provided gene lists in the markers of clusters in the
seurat object. All genes present in the Seurat object are used as the background
gene set. This function is called by find_write_markers, but can also
be used independently.
}
\examples{
\dontrun{
# This is better with longer gene lists, but this is just an example
my_genes <- list(t_cell = c("Cd3d", "Cd4", "Cd8"),
                 beta   = c("Ins1", "Ins2", "Nkx2-2"),
                 alpha  = c("Gcg", "Arx"))
DE_res <- FindMarkers(seurat_object, only.pos = TRUE)
enrichment <- hypergeometric_test(seurat_object = seurat_object,
                                  gene_list     = my_list,
                                  DE_table      = DE_res)
}
}
