% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_seurat_v5}
\alias{is_seurat_v5}
\title{Check Seurat version}
\usage{
is_seurat_v5()
}
\value{
Logical value indicating if SeuratObject version is 5.0.0 or higher.
}
\description{
Determines if the SeuratObject package version is 5.0.0 or higher.
}
