% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{make_go_plot_single}
\alias{make_go_plot_single}
\title{Makes go plots of different terms for one DE test}
\usage{
make_go_plot_single(
  gost_output,
  gost_query,
  gost_source,
  plot_colors = c("blue", "red"),
  intersection_cutoff = 5
)
}
\arguments{
\item{gost_output}{the output from running gost.}

\item{gost_query}{the name of the cell type or sample from DE to plot}

\item{gost_source}{what source to plot. Can be any sources returned by gost}

\item{plot_colors}{OPTIONAL colors for the samples. Default is blue to red.}

\item{intersection_cutoff}{OPTIONAL how many intersecting genes there must
be to plot. Default is 5.}
}
\description{
This function is meant to be called by make_go_plots and not used on it's own!!!
This function makes plots based on different go terms after running gost.
This function is primarily meant to be called by other functions.
}
\keyword{internal}
