% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{make_go_plots}
\alias{make_go_plots}
\title{Makes go plots}
\usage{
make_go_plots(
  gost_output,
  sources = c("GO:BP", "KEGG", "REAC", "TF"),
  plot_colors = c("blue", "red"),
  intersection_cutoff = 5
)
}
\arguments{
\item{gost_output}{the output from running gost.}

\item{sources}{OPTIONAL a list of what sources to plot. Can be any sources
returned by gost. Default is GO:BP, KEGG, REAC and TF}

\item{plot_colors}{OPTIONAL colors for the samples. Default is blue to red.}

\item{intersection_cutoff}{OPTIONAL how many intersecting genes there must
be to plot. Default is 5.}
}
\description{
This function is meant to be called by run_gost and not used on it's own!!!
This function makes plots based on different go terms after running gost.
This function is primarily meant to be called by other functions.
}
\keyword{internal}
