% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_clusters.R
\name{name_clusters}
\alias{name_clusters}
\title{Name clusters based on a reference}
\usage{
name_clusters(
  seurat_object,
  ref_mat,
  save_dir,
  save_name,
  ADT = FALSE,
  assay = "RNA",
  nfeatures = 1000,
  clusters = "RNA_cluster",
  plot_type = "rna.umap",
  cor_cutoff = 0.5,
  features = NULL
)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{ref_mat}{A matrix from a reference. This should be genes by cell type. If it
comes from a single cell dataset, you can make this reference using average_clusters
from clustifyr. See the clustifyr tutorial for more information.}

\item{save_dir}{A path to a directory to save the output matricies and plots.}

\item{save_name}{The name to use to save the cell type information. Example is
t_cell_ref. The returned seurat object will have a column name RNA_t_cell_ref
in the meta data.}

\item{ADT}{OPTIONAL If you want to also name cell types based on clusters determined
based on ADTs. Only recommened if you used a large pannel of ADTs. To use this,
PCA_dimRed and group_cells must have been run on the ADT assay. Additionally, a combined
cluster must also have been generated. The steps to do this will eventually be added to
this package, but they have not yet.}

\item{assay}{OPTIONAL What assay to use to compare to the reference. "RNA" is
receommended. Default is "RNA"}

\item{nfeatures}{OPTIONAL The number of features to use to run clustifyr. Check the
clustifyr tutorial for a recommendation. Normally 1000 is a good place to start. Too
high of a value will lead to many of the cell types having a similar correlation.
Default is 1000.}

\item{clusters}{OPTIONAL What existing clusters to use from your seurat object to identify
cell types with clustifyr. Can be any discrete column from your metadata. Default is
"RNA_cluster"}

\item{plot_type}{OPTIONAL What plot type should be used to plot the results. These plots
will be saved in the provided directory in a subdirectory "images". Can be any plot
type in the seurat object. Default is "rna.umap".}

\item{cor_cutoff}{OPTIONAL The correlation value to use to determine if a cell type
is named. If no cell types have a correlation above this value for a given cluster,
that cluster will be named "undetermined" in the final output. Default is 0.5}

\item{features}{OPTIONAL A list of features to use instead of running FindVariableFeatures
Default is NULL. If set, will override the nfeatures argument. This is helpful if you
specifically want to exclude features (for example Ig genes or TCR genes).}
}
\value{
a list containing the correlation matricies and the seurat object with a new
meta data column including the identified cell type.
}
\description{
This function uses clustifyr to name clusters based on a reference dataset.
}
\examples{
\dontrun{
clustify_list <- name_clusters(seurat_object = seurat_object,
                               ref_mat       = t_cell_matrix,
                               save_dir      = "clustifyr",
                               save_name     = t_cell_ref)
seurat_object <- clustify_list$object
}
}
