% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{pairwise_markers}
\alias{pairwise_markers}
\title{Run pairwise differential expression}
\usage{
pairwise_markers(seurat_object, meta_col, p_val_cutoff = 0.05, ...)
}
\arguments{
\item{seurat_object}{A seurat object that has already been normalized.}

\item{meta_col}{The column in the meta data to use to identify all
pairwise markers}

\item{p_val_cutoff}{OPTIONAL The cutoff for p value for returned genes.
Default is 0.05}

\item{...}{OPTIONAL extra arguments passed to FindMarkers}
}
\value{
A dataframe containing all pairwise markers.
}
\description{
This function runs  pairwise differential expression analysis on all
possible combinations in the meta data column provided using seurat's
FindMarkers
}
\examples{
\dontrun{
marker_df <- pairwise_markers(seurat_object = seurat_object,
                              meta_col      = "clusters")
}
}
