% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{PCA_dimRed}
\alias{PCA_dimRed}
\title{Performs PCA dimensionality reduction}
\usage{
PCA_dimRed(
  sample_object,
  assay = "RNA",
  reduction_name = NULL,
  vars_to_regress = NULL
)
}
\arguments{
\item{sample_object}{A seurat object that has already been normalized}

\item{assay}{OPTIONAL The name of the assay to compute PCA for. Can be "RNA",
"ADT", "SCT", or "integrated". Default is "RNA"}

\item{reduction_name}{OPTIONAL The name to give the output PCA reduction. Defaults
are "pca" (for RNA assay), "apca" (for ADT assay), "sctpca" (for SCT assay), and
"integrated" (for integrated assay)}

\item{vars_to_regress}{OPTIONAL What variables to regress out when scaling the data.
Only used if assay = "ADT" (otherwise the data should already be scaled). Deefault
is NULL}
}
\value{
A seurat object with an added reduction named based on the reduction_name
given or based on the rules described above.
}
\description{
This function will perform PCA dimensionality reduction and create a reduction
name for RNA, SCT, ADT, or integrated assays. There is currently no functionality
included for other assays.
}
\examples{
splen_so <- PCA_dimRed(sample_object = splen_so)
splen_so <- SCTransform(splen_so, verbose = FALSE)
splen_so <- PCA_dimRed(sample_object = splen_so,
                       assay         = "SCT")
}
