% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(
  seurat_object,
  gene_list,
  meta_col,
  colors = NULL,
  meta_df = NULL,
  color_list = NULL,
  max_val = 2.5,
  min_val = -2.5,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  average_expression = FALSE,
  plot_meta_col = TRUE,
  plot_rownames = TRUE,
  cell_order = NULL,
  ...
)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{gene_list}{What genes to use in the heatmap. The function will take a long time
and the heatmap will be huge if you use all genes.}

\item{meta_col}{What column from the meta data should be used to annotate the x axis
of the heatmap.}

\item{colors}{OPTIONAL What colors should be used to annotate the x axis. Default is
Set1 from RColorBrewer}

\item{meta_df}{OPTIONAL If you want more complex annotations than just annotating by
one column of the meta data, you can include your own meta_df here. If you use this
option, you must include your own color list, where the names of the list match
each column of your meta_df and the vector of colors in each item of the list match
the number of items in the meta data column. Additionally, if you use this
option, you must make sure that the cell names from the meta data match the cell names
in your Seurat object. I've found that this is a particular issue when using average
expression of clusters.}

\item{color_list}{OPTIONAL only needs to be included when meta_df is set. This should be
a list where the names of the list match each column of your meta_df and the vector of
colors in each item of the list match the number of items in the meta data column.}

\item{max_val}{OPTIONAL What value to cap the z-score shown. This makes it so highly
variable genes don't completely controll the color scale. Default is 2.5}

\item{min_val}{OPTIONAL What value to cap the z-score shown. This makes it so highly
variable genes don't completely controll the color scale. Default is -2.5}

\item{cluster_rows}{OPTIONAL if rows should be clustered. Default is FALSE}

\item{average_expression}{OPTIONAL if the average expression of clusters should be
computed before making the heatmap. This option makes much prettier plots, but
can be a bit buggy, I'm still working on making it work more robustly.}

\item{...}{Other options passed to \code{pheatmap}}

\item{cluster_columns}{OPTIONAL if columns should be clustered. Default is FALSE}

\item{show_rownames}{OPTIONAL if rownames should be printed. Default is TRUE. I generally
like to print the rownames unless there are too many to be visually pleasng.}
}
\value{
A pheatmap object with x axis colors based on the meta data column provided
or your own meta_df with color_list.
}
\description{
This function will plot a heatmap with annotations based on the metadata from a
supplied seurat object. The heatmap is colored by the blueYellow palette from
the archR package.
}
\examples{
\dontrun{
plot_heatmap(seurat_object = seurat_object,
             gene_list     = c("INS1", "INS2", "GCG", "SST", "NKX2-2"),
             meta_col      = "clusters",
             color         = c(1 = "#FFFFFF", 2 = "#FF0000"))
plot_heatmap(seurat_object      = seurat_object,
             gene_list          = c("INS1", "INS2", "GCG", "SST", "NKX2-2"),
             meta_col           = "clusters",
             color              = c(1 = "#FFFFFF", 2 = "#FF0000"),
             average_expression = TRUE)
#' plot_heatmap(seurat_object = seurat_object,
             gene_list     = gene_set,
             meta_col      = "clusters")
}
}
