% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDF}
\alias{plotDF}
\title{Make a dataframe}
\usage{
plotDF(seurat_object, y_val, x_val, col_by = NULL, assay = NULL)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{y_val}{What to use to create the y-axis}

\item{x_val}{What to use to separate along the x-axis}

\item{col_by}{OPTIONAL what to use to color the cells (or violins).}
}
\description{
This function is meant to be called by featDistPlot and not used on it's own!!!
This function will make a dataframe to pass to plotting functions
}
\keyword{internal}
