% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot a group of dimensionality reductions}
\usage{
plotDimRed(
  sample_object,
  col_by,
  save_plot = NULL,
  plot_type = "umap",
  dims_use = NULL,
  highlight_group = FALSE,
  group = NULL,
  meta_data_col = "orig.ident",
  ...
)
}
\arguments{
\item{sample_object}{A seurat object}

\item{col_by}{What to color the plot by. May be a gene, adt, or column
from the metadata. It can either be discrete or continuous. This can be a vector
of selections (example c("cluster", "Ins1")). If a vector is given, the plots will
be returned as a list in the order of the supplied vector.}

\item{save_plot}{OPTIONAL A path to a file name if you want the plot saved. It
will be saved as a PDF. If nothing is provided, the plot will be returned but not
saved}

\item{plot_type}{OPTIONAL The type of plot. The possible plots can be found in
"reductions" in your seurat object. You can find these with names(object@reductions)
Default is "umap".}

\item{dims_use}{OPTIONAL what dimensions to plot. Default is to plot dimensions 1 and 2}

\item{highlight_group}{OPTIONAL if only some cells should be highlighted on the plot.
when set to TRUE you must also set group and meta_data_col. When TRUE, the selected
group will be colored and all other cells will be grey.}

\item{group}{The group to color. This group must be present in the selected meta data
column.}

\item{meta_data_col}{The column in the seurat metadata (options can be found by calling
colnames(object[[]])) used for coloring the desired group. For example, if you want to
only highlight one cluster, meta_data_col = "cluster".}

\item{...}{OPTIONAL arguments supplied to plotDimRedSingle}
}
\value{
a list of plots colored by the given parameteres
}
\description{
This function will plot a dimensionality reduction that exists in your seurat
object (ex. UMAP or PCA). You can color the plot using any gene, adt (if they
exist) or meta data column. You can also highlight just one group from an
existing meta data column. This function will accept a vector in the col_by
argument. These can be any combination of genes or meta data columns. If
multiple options are given, a list of plots will be returned.
}
\examples{
\dontrun{
plotDimRed(sample_object = seurat_object,
           col_by        = "clusters",
           color         = c(1 = "#FFFFFF", 2 = "#FF0000"))
plotDimRed(sample_object = seurat_object,
           col_by        = "INS1")
plotDimRed(sample_object   = seurat_object,
           col_by          = "INS1",
           highlight_group = TRUE,
           group           = 1,
           meta_data_col   = "clusters")
}
}
