% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDimRedSingle}
\alias{plotDimRedSingle}
\title{Plot one dimensionality reduction}
\usage{
plotDimRedSingle(
  seurat_object,
  col_by,
  plot_type = "umap",
  dims_use = NULL,
  highlight_group = FALSE,
  group = NULL,
  meta_data_col = "orig.ident",
  assay = NULL,
  reorder_cells = FALSE,
  ...
)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{col_by}{What to color the plot by. May be a gene, adt, or column
from the metadata. It can either be discrete or continuous.}

\item{plot_type}{OPTIONAL The type of plot. The possible plots can be found in
"reductions" in your seurat object. You can find these with names(object@reductions)
Default is "umap".}

\item{dims_use}{OPTIONAL what dimensions to plot. Default is to plot dimensions 1 and 2}

\item{highlight_group}{OPTIONAL if only some cells should be highlighted on the plot.
when set to TRUE you must also set group and meta_data_col. When TRUE, the selected
group will be colored and all other cells will be grey.}

\item{group}{OPTIONAL The group to color. This group must be present in the selected
meta data column.}

\item{meta_data_col}{The column in the seurat metadata (options can be found by calling
colnames(object[[]])) used for coloring the desired group. For example, if you want to
only highlight one cluster, meta_data_col = "cluster".}

\item{assay}{OPTIONAL The assay to use to color the plot. This is mostly useful if
you have the same names in your ADT and RNA slots so you can ensure you plot the
correct one.}

\item{reorder_cells}{OPTIONAL If cells should be randomly reodered to improve plotting
aesthetics. Can be helpful if you don't want samples plotted in order. Default is FALSE.}

\item{...}{OPTIONAL arguments supplied to groupDiscretePlots, groupContinuousPlots,
discretePlots, or continuousPlots}
}
\description{
This function is meant to be called by plotDimRed and not used on it's own!!!
This function will plot a dimensionality reduction that exists in your seurat
object (ex. UMAP or PCA). You can color the plot using any gene, adt (if they
exist) or meta data column. You can also highlight just one group from an
existing meta data column.
}
\keyword{internal}
