% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{run_gost}
\alias{run_gost}
\title{Run gost gene ontology}
\usage{
run_gost(
  seurat_de = NULL,
  seurat_object = NULL,
  organism = "mmusculus",
  sources = c("GO:BP", "KEGG", "REAC", "TF"),
  plot_colors = c("blue", "red"),
  intersection_cutoff = 5,
  pval = 0.05,
  correction_method = "fdr",
  ...
)
}
\arguments{
\item{seurat_de}{OPTIONAL A de data frame returned by find_markers.
Must provide either a seurat object or seurat de}

\item{seurat_object}{OPTIONAL A seurat object that has already been
normalized. Must provide either a seurat object or a seurat de.}

\item{organism}{OPTIONAL What organism to use. Will be past to gprofiler2::gost}

\item{sources}{OPTIONAL a list of what sources to plot. Can be any sources
returned by gost. Default is GO:BP, KEGG, REAC and TF}

\item{plot_colors}{OPTIONAL colors for the samples. Default is blue to red.}

\item{intersection_cutoff}{OPTIONAL how many intersecting genes there must
be to plot. Default is 5.}

\item{pval}{OPTIONAL the pvalue to use as a cutoff, default is 0.05}

\item{correction_method}{OPTIONAL what correction method to use. Default is 'fdr'}

\item{...}{Other arguments past to gprofiler2::gost}
}
\value{
A list containing two items, gost_output and go_plots
}
\description{
This function allows you to run gost given a matrix of DE output or seurat
object
}
\examples{
\dontrun{
go_list <- run_gost(seurat_object = seurat_object)
go_list <- run_gost(seurat_de = de_df)
}
}
