% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{run_pathview}
\alias{run_pathview}
\title{Run pathview on single cell data}
\usage{
run_pathview(
  gene_matrix,
  path_id,
  path_name = NULL,
  out_dir = "pathways",
  multi_state = FALSE,
  gen_id = "mmu"
)
}
\arguments{
\item{gene_matrix}{a matrix of log fold change values. Columns are samples
and rownames are genes. Can include NA values. Must be in a matrix format.}

\item{path_id}{the KEGG 5 number id for a pathway.}

\item{path_name}{OPTIONAL the name of the KEGG pathway. This is used to
name the output files. If this isn't set, output files are named based
on the path id}

\item{out_dir}{OPTIONAL the output directory to copy the files into. This
directory will be created if it doesn't already exist. Default is "pathways"}

\item{multi_state}{OPTIONAL if multiple samples are included should they
all be plotted on the same output image. Default is FALSE}

\item{gen_id}{OPTIONAL the species ID. Default is "mmu" for mouse. Set if
you aren't using mouse}
}
\description{
This function is meant to be called by de_to_pathview and not used on it's own!!!
This function allows you to run pathview given a matrix of DE output
}
\keyword{internal}
