% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{save_gost}
\alias{save_gost}
\title{Save output from running gost}
\usage{
save_gost(
  gost_output,
  save_dir_plots,
  save_dir_text = NULL,
  save_plots = TRUE,
  save_text = TRUE,
  save_excel = TRUE
)
}
\arguments{
\item{gost_output}{the output from running run_gost function.}

\item{save_dir_plots}{the directory to save the plots to}

\item{save_dir_text}{OPTIONAL the directory to save the text to. If left
NULL, will default to the same directory as the plots}

\item{save_plots}{OPTIONAL if plots should be saved. Default is TRUE}

\item{save_text}{OPTIONAL if text should be saved to a csv file. The whole
gost output will be saved. Default is TRUE.}

\item{save_excel}{OPTIONAL if an excel file should be created. A file will
be created for each query and a tab will be inserted for each type of output.
Requres openxlsx is installed.}
}
\value{
No return, writes either a csv or excel workbook of gene onotology results
and saves plots to a pdf
}
\description{
This function saves both plots and text output from running gost
}
\examples{
\dontrun{
save_gost(gost_output    = gost_output,
          save_dir_plots = "plots")
save_gost(gost_output    = gost_output,
          save_dir_plots = "plots",
          save_dir_text  = "files",
          save_excel     = FALSE)
}
}
