% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnalysis.R
\docType{package}
\name{scAnalysisR-package}
\alias{scAnalysisR-package}
\alias{_PACKAGE}
\alias{scAnalysisR}
\title{scAnalysisR: Single Cell Analysis Pipeline}
\description{
Packages to aid in the analysis of single cell RNA-seq data. Many functions are wrappers around seurat and ggplot. Requires seurat objects.
}
\author{
\strong{Maintainer}: Kristen Wells \email{kristen.wells-wrasman@cuanschutz.edu} (\href{https://orcid.org/0000-0002-7466-8164}{ORCID})

}
\keyword{internal}
