% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_seurat_assay}
\alias{set_seurat_assay}
\title{Set Seurat Assay}
\usage{
set_seurat_assay(object, new_data, type = "counts", assay = "RNA")
}
\arguments{
\item{object}{A seurat object}

\item{new_data}{What new data should be added}

\item{type}{What type should be updated, data or counts}

\item{assay}{What assay the data should be added to}
}
\value{
The resulting data matrix.
}
\description{
Determines the seurat version and grabs the assay data using
the appropriate arguments.
}
