% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{violinPlot}
\alias{violinPlot}
\title{Plot a violin plot}
\usage{
violinPlot(
  seurat_object,
  y_val,
  x_val,
  col_by = NULL,
  color = NULL,
  plot_jitter = FALSE,
  plot_median = TRUE,
  dodge = 1,
  width = 0.9,
  assay = NULL
)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{y_val}{What to use to create the y-axis}

\item{x_val}{What to use to separate along the x-axis}

\item{col_by}{OPTIONAL what to use to color the cells (or violins).}

\item{color}{OPTIONAL The color palette used to color. Default is Set1 from RColorBrewer}

\item{plot_jitter}{OPTIONAL if a jitter plot should also be made. Default is FALSE.}

\item{plot_median}{OPTIONAL if the median value should be included in the
violin plot. Default is TRUE}

\item{dodge}{OPTIONAL how to adjust the placement of the violin plot. Default is 1}

\item{width}{OPTIONAL how to adjust how wide the violin plots are. Default is 0.9}
}
\description{
This function is meant to be called by featDistPlot and not used on it's own!!!
This function will plot a violin plot.
}
\keyword{internal}
