% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{create_seurat_object}
\alias{create_seurat_object}
\title{Creates a seurat object}
\usage{
create_seurat_object(
  sample,
  count_path,
  ADT = TRUE,
  hashtag = TRUE,
  min_features = 200,
  min_cells = 3,
  hashtag_idents = NULL,
  tenx_structure = "multi"
)
}
\arguments{
\item{sample}{The name of the sample (also the name of the directory containing the
count matricies).}

\item{count_path}{The full path to the 10x output files. In total this will be
count_path/sample/outs/count/filtered_feature_bc_matrix}

\item{ADT}{OPTIONAL If ADTs were included in the experiment. Default is TRUE}

\item{hashtag}{OPTIONAL If HTOs were included in the experiment. Default is TRUE}

\item{min_features}{OPTIONAL How many features per cell are necessary to keep the cell
will be passed to CreateSeuratObject. Default is 200}

\item{min_cells}{OPTIONAL the number of cells required to keep a feature, will be
passed to CreateseuratObject. Default is 3.}

\item{hashtag_idents}{OPTIONAL the identities of the hashtags. If provided must be
exactly the identity that would be in the rowname. Default is to search for the
word "Hashtag" in the identity.}

\item{tenx_structure}{OPTIONAL if the file structure is from 10x multi (sample/outs/count),
10x count (sample/outs/) 10x multi 7 (sample/outs/multi/count or neither (sample).
Options are "multi" (default), "count", "multi7", or "none"}
}
\value{
A seurat object with assays for HTO and ADT (if HTO and ADT are true).
If HTO and ADTs are included, those matricies will be normalized by CLR normalization
}
\description{
This function will create a seurat object. It assumes that the 10x output files are
in a directory named by the sample name followe by outs/count/filtered_feature_bc_matrix
This function can also generate matricies for HTO and ADTs. This function will also
perform CLR normlization on the ADT and HTO matricies, but will not normalize the
RNA data.
}
\examples{
\dontrun{
seurat_object <- create_seurat_object(sample     = "WT_1",
                                      count_path = "/home/users/kwells/experiment_1/")
seurat_object <- create_seurat_object(sample     = "WT_1",
                                      count_path = "/home/users/kwells/experiment_2/",
                                      ADT        = FALSE,
                                      hashtag    = FALSE)
}
}
