% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_average_expression}
\alias{get_average_expression}
\title{Get average expression}
\usage{
get_average_expression(object, cell_type_col = "cell_type", assay = "RNA", ...)
}
\arguments{
\item{object}{A seurat object}

\item{cell_type_col}{The name of column containing cell type information.
Default is \code{cell_type}.}

\item{assay}{What assay the data should be pulled from}

\item{...}{arguments passed to either AverageExpression (Seurat v4)
or PseudobulkExpression (seurat V5)}
}
\value{
A data frame of the average expression of the gene.
}
\description{
Helper function to calculate and label the average expression
}
